// frontend/sdk/models/AdminUserDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sopranino API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AdminUserDto
 */
export interface AdminUserDto {
    /**
     * 
     * @type {string}
     * @memberof AdminUserDto
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof AdminUserDto
     */
    spotifyUserId?: string;
    /**
     * 
     * @type {string}
     * @memberof AdminUserDto
     */
    displayName: string;
    /**
     * 
     * @type {string}
     * @memberof AdminUserDto
     */
    avatarUrl?: string;
    /**
     * 
     * @type {boolean}
     * @memberof AdminUserDto
     */
    isTrusted: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof AdminUserDto
     */
    isAdmin: boolean;
    /**
     * 
     * @type {Date}
     * @memberof AdminUserDto
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof AdminUserDto
     */
    updatedAt: Date;
}

/**
 * Check if a given object implements the AdminUserDto interface.
 */
export function instanceOfAdminUserDto(value: object): value is AdminUserDto {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('displayName' in value) || value['displayName'] === undefined) return false;
    if (!('isTrusted' in value) || value['isTrusted'] === undefined) return false;
    if (!('isAdmin' in value) || value['isAdmin'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    if (!('updatedAt' in value) || value['updatedAt'] === undefined) return false;
    return true;
}

export function AdminUserDtoFromJSON(json: any): AdminUserDto {
    return AdminUserDtoFromJSONTyped(json, false);
}

export function AdminUserDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): AdminUserDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'spotifyUserId': json['spotifyUserId'] == null ? undefined : json['spotifyUserId'],
        'displayName': json['displayName'],
        'avatarUrl': json['avatarUrl'] == null ? undefined : json['avatarUrl'],
        'isTrusted': json['isTrusted'],
        'isAdmin': json['isAdmin'],
        'createdAt': (new Date(json['createdAt'])),
        'updatedAt': (new Date(json['updatedAt'])),
    };
}

export function AdminUserDtoToJSON(json: any): AdminUserDto {
    return AdminUserDtoToJSONTyped(json, false);
}

export function AdminUserDtoToJSONTyped(value?: AdminUserDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'spotifyUserId': value['spotifyUserId'],
        'displayName': value['displayName'],
        'avatarUrl': value['avatarUrl'],
        'isTrusted': value['isTrusted'],
        'isAdmin': value['isAdmin'],
        'createdAt': value['createdAt'].toISOString(),
        'updatedAt': value['updatedAt'].toISOString(),
    };
}